backend/requirements.txt
# FastAPI & Server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database (using lighter supabase client)
supabase==2.4.0
python-dotenv>=1.0.0

# Resume Parsing
PyMuPDF>=1.24.0
python-docx>=1.1.0

# AI / Ollama
ollama>=0.3.0
httpx>=0.27.0

# Utilities
pydantic>=2.9.0
pydantic-settings>=2.5.0
